add fragment to show book list
