+++ app/src/main/java/com/example/abhishek/bookshareapp/ui/SearchResultsActivity.java
@@ -2,26 +2,18 @@ package com.example.abhishek.bookshareapp.ui;
 
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
-import android.util.Log;
 import android.view.View;
 import android.widget.EditText;
 import android.widget.ListView;
-import android.widget.Toast;
 
 import com.example.abhishek.bookshareapp.R;
-import com.example.abhishek.bookshareapp.api.BooksAPI;
-import com.example.abhishek.bookshareapp.api.NetworkingFactory;
 import com.example.abhishek.bookshareapp.api.models.Book;
-import com.example.abhishek.bookshareapp.api.models.GoodreadsResponse;
 import com.example.abhishek.bookshareapp.api.models.Search;
+import com.example.abhishek.bookshareapp.ui.adapter.BookListFragment;
 import com.example.abhishek.bookshareapp.ui.adapter.BooksAdapter;
 import com.example.abhishek.bookshareapp.utils.CommonUtilities;
 
 import java.util.List;
-
-import retrofit2.Call;
-import retrofit2.Callback;
-import retrofit2.Response;
 
 /**
  * Created by abhishek on 13/2/16.
@@ -35,53 +27,27 @@ public class SearchResultsActivity extends AppCompatActivity {
     String API_KEY= CommonUtilities.API_KEY;
     Search sr;
     EditText searchEditText;
+    BookListFragment bookListFragment;
+
     @Override
     public void onCreate(Bundle savedInstanceState) {
 
         super.onCreate(savedInstanceState);
         setContentView(R.layout.search_results);
 
-        resultsList= (ListView)findViewById(R.id.results_list);
         searchEditText = (EditText) findViewById(R.id.searchEditText);
+
+        bookListFragment = new BookListFragment();
+
+        getFragmentManager().beginTransaction()
+                .replace(R.id.container, bookListFragment)
+                .commit();
 
     }
 
     public void search(View view){
-
         query = searchEditText.getText().toString();
-        getBooks(query, API_KEY);
-
-    }
-
-    public void getBooks(String query,String key) {
-
-        Toast.makeText(SearchResultsActivity.this,"getbooks",Toast.LENGTH_SHORT).show();
-
-        BooksAPI api = NetworkingFactory.getInstance().getBooksApi();
-        Call<GoodreadsResponse> call = api.getBooks(query, key);
-        call.enqueue(new Callback<GoodreadsResponse>() {
-            @Override
-            public void onResponse(Call<GoodreadsResponse> call, Response<GoodreadsResponse> response) {
-                if(response.body()!=null) {
-                    Log.d("searchresp", response.toString());
-                    sr = response.body().getSearch();
-                    bookList = sr.getBooks();
-                    Log.d("searchresp", bookList.toString());
-                    adapter = new BooksAdapter(SearchResultsActivity.this, bookList);
-                    resultsList.setAdapter(adapter);
-
-                    Toast.makeText(SearchResultsActivity.this, sr.getBooks().get(0).getBookDetails().getAuthor().getAuthor_name(), Toast.LENGTH_SHORT).show();
-                }
-
-            }
-
-            @Override
-            public void onFailure(Call<GoodreadsResponse> call, Throwable t) {
-                Log.d("searchresp","searchOnFail "+ t.toString());
-
-            }
-        });
-
+        bookListFragment.getBooks(query, API_KEY);
     }
 
 }
