+++ app/src/main/res/layout/search_results.xml
@@ -21,9 +21,10 @@
             android:onClick="search"/>
 
     </LinearLayout>
-    <ListView
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:id="@+id/results_list"></ListView>
+
+    <FrameLayout
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:id="@+id/container"/>
 
 </LinearLayout>
